Eased work across servers by re-working code.
I reworked most of the infrastructure around the bot to work across many servers. (In layman's terms: now it can work in more servers than before)
I also added some more commands to allow for BreenBot to modify people's roles, such as grant and revoke.
Now, the server that send the ISLog() function will also be the one to receive the output.

The code grew about 50 lines bigger in just this one commit!